+++ packages/sitecore-jss-angular/src/public_api.ts
@@ -42,6 +42,7 @@ export {
 } from '@sitecore-jss/sitecore-jss/layout';
 export { RetryStrategy, DefaultRetryStrategy } from '@sitecore-jss/sitecore-jss/graphql';
 export { constants, HttpDataFetcher, HttpResponse } from '@sitecore-jss/sitecore-jss';
+export { ClientError } from 'graphql-request';
 export {
   isServer,
   isExperienceEditorActive,