+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -5,6 +5,7 @@ export {
   HttpResponse,
   AxiosDataFetcher,
   AxiosDataFetcherConfig,
+  ClientError,
 } from '@sitecore-jss/sitecore-jss';
 export {
   isEditorActive,